+++ include/global-types.h
@@ -1,0 +1,9 @@
+#pragma once
+#include <cstdint>
+
+using u8 = std::uint8_t;
+using u16 = std::uint16_t;
+using u32 = std::uint32_t;
+using u64 = std::uint64_t;
+using s8 = std::int8_t;
+using s16 = std::int16_t;
