+++ scripts/test.sh
@@ -38,5 +38,5 @@ else
 fi
 
 # Example usage:
-# ./scripts/test.sh "CPUTestFixture.IMM"   # Run a specific test
+# ./scripts/test.sh "CpuTest.IMM"   # Run a specific test
 # ./scripts/test.sh                      # Run all tests in parallel
