+++ tests/cpu_test.cpp
@@ -1,69 +1,4 @@
-// cpu_test.cpp
-#include "bus.h"
-#include "cpu.h"
-#include "ppu.h"
-#include "cartridge.h"
-#include "json.hpp"
-#include <cstdint>
-#include <fstream>
-#include <gtest/gtest.h>
-#include <iomanip>
-#include <ios>
-#include <iostream>
-#include <sstream>
-#include <stdexcept>
-#include <string>
-#include "config.h"
-
-using json = nlohmann::json;
-
-// forward declarations
-void printTestStartMsg( const std::string &testName );
-void printTestEndMsg( const std::string &testName );
-json extractTestsFromJson( const std::string &path );
-
-class CPUTestFixture : public ::testing::Test
-// This class is a test fixture that provides shared setup and teardown for all tests
-{
-  protected:
-    // All tests assume flat memory model, which is why true is passed to Bus constructor
-    Bus  bus;
-    CPU &cpu = bus.cpu;
-    PPU &ppu = bus.ppu;
-
-    CPUTestFixture() = default;
-
-    void                      RunTestCase( const json &testCase );
-    void                      LoadStateFromJson( const json &jsonData, const std::string &state );
-    [[nodiscard]] std::string GetCPUStateString( const json &jsonData, const std::string &state ) const;
-
-    // Expose private methods
-    // CPUTestFixture is a friend class of CPU. To use cpu private methods, we need to create
-    // wrappers.
-    void               SetFlags( u8 flag );
-    void               ClearFlags( u8 flag );
-    [[nodiscard]] bool IsFlagSet( u8 flag ) const;
-    [[nodiscard]] u8   Read( u16 address ) const;
-    void               Write( u16 address, u8 data ) const;
-    u16                IMM();
-    u16                ZPG();
-    u16                ZPGX();
-    u16                ZPGY();
-    u16                ABS();
-    u16                ABSX();
-    u16                ABSY();
-    u16                IND();
-    u16                INDX();
-    u16                INDY();
-    u16                REL();
-
-    void LoadTestCartridge()
-    {
-        std::string romFile = std::string( ROM_DIR ) + "/palette.nes";
-        bus.cartridge.LoadRom( romFile );
-        bus.cpu.Reset();
-    }
-};
+#include "cpu_test.h"
 
 /*
 ################################################################
@@ -72,23 +7,23 @@ class CPUTestFixture : public ::testing::Test
 ||                                                            ||
 ################################################################
  */
-TEST_F( CPUTestFixture, SanityCheck )
+TEST_F( CpuTest, SanityCheck )
 {
     // cpu.read and cpu.write shouldn't throw any errors
-    u8 const testVal = Read( 0x0000 );
-    Write( 0x0000, testVal );
+    u8 const testVal = cpu.Read( 0x0000 );
+    cpu.Write( 0x0000, testVal );
 }
 
-TEST_F( CPUTestFixture, MemorySweep )
+TEST_F( CpuTest, MemorySweep )
 {
     // Write to every place in addressable memory, ensure emulator doesn't crash
     try {
         for ( u16 i = 0; i < 0xFFFF; ++i ) {
-            Write( i, 0x00 );
+            cpu.Write( i, 0x00 );
         }
 
         for ( u16 i = 0; i < 0xFFFF; ++i ) {
-            auto dummyVal = Read( i );
+            auto dummyVal = cpu.Read( i );
             (void) dummyVal;
         }
     } catch ( std::exception &e ) {
@@ -99,11 +34,11 @@ TEST_F( CPUTestFixture, MemorySweep )
 
     try {
         for ( u16 i = 0; i < 0xFFFF; ++i ) {
-            Write( i, 0x00 );
+            cpu.Write( i, 0x00 );
         }
 
         for ( u16 i = 0; i < 0xFFFF; ++i ) {
-            auto dummyVal = Read( i );
+            auto dummyVal = cpu.Read( i );
             (void) dummyVal;
         }
     } catch ( std::exception &e ) {
@@ -111,27 +46,27 @@ TEST_F( CPUTestFixture, MemorySweep )
     }
 }
 
-TEST_F( CPUTestFixture, RamCheck )
+TEST_F( CpuTest, RamCheck )
 {
     // Write 1 to every location in RAM
     for ( u16 i = 0x0000; i < 0x0800; ++i ) {
-        Write( i, 0x01 );
+        cpu.Write( i, 0x01 );
     }
     for ( u16 i = 0x0000; i < 0x0800; ++i ) {
-        EXPECT_EQ( Read( i ), 0x01 );
+        EXPECT_EQ( cpu.Read( i ), 0x01 );
     }
 
     bus.DebugReset();
     // Write 1 to mirrored locations in RAM
     for ( u16 i = 0x0800; i < 0x2000; ++i ) {
-        Write( i, 0x01 );
+        cpu.Write( i, 0x01 );
     }
     for ( u16 i = 0x0000; i < 0x0800; ++i ) {
-        EXPECT_EQ( Read( i ), 0x01 );
+        EXPECT_EQ( cpu.Read( i ), 0x01 );
     }
 }
 
-TEST_F( CPUTestFixture, ResetVector )
+TEST_F( CpuTest, ResetVector )
 {
     bus.EnableJsonTestMode(); // enables flat memory
     cpu.Write( 0xFFFC, 0x34 );
@@ -145,7 +80,7 @@ TEST_F( CPUTestFixture, ResetVector )
     EXPECT_EQ( cpu.GetProgramCounter(), 0x1234 );
 }
 
-TEST_F( CPUTestFixture, IRQ )
+TEST_F( CpuTest, IRQ )
 {
     bus.EnableJsonTestMode();
     bus.DebugReset();
@@ -175,7 +110,7 @@ TEST_F( CPUTestFixture, IRQ )
     EXPECT_EQ( cpu.GetProgramCounter(), 0x1234 );
 }
 
-TEST_F( CPUTestFixture, NMI )
+TEST_F( CpuTest, NMI )
 {
     bus.EnableJsonTestMode();
     bus.DebugReset();
@@ -195,7 +130,7 @@ TEST_F( CPUTestFixture, NMI )
     EXPECT_EQ( cpu.GetProgramCounter(), 0x1234 );
 }
 
-TEST_F( CPUTestFixture, ExecuteFrame )
+TEST_F( CpuTest, ExecuteFrame )
 {
 }
 
@@ -209,50 +144,24 @@ TEST_F( CPUTestFixture, ExecuteFrame )
 
 /* This is a macro to simplify test creation for json tests
  * The naming convention is <opcode hex>_<mnemonic>_<addressing mode>
- * e.g. x00_BRK_Implied, x01_ORA_IndirectX, x05_ORA_ZeroPage, etc.
- */
-// NOLINTBEGIN(cppcoreguidelines-macro-usage)
+ * e.g. x00_BRK_Implied, x01_ORA_IndirectX, x05_ORA_ZeroPage, etc. */
 #define CPU_TEST( opcode_hex, mnemonic, addr_mode, filename )                                                \
-    TEST_F( CPUTestFixture, x##opcode_hex##_##mnemonic##_##addr_mode )                                       \
+    TEST_F( CpuTest, x##opcode_hex##_##mnemonic##_##addr_mode )                                              \
     {                                                                                                        \
-        std::string const testName = #opcode_hex " " #mnemonic " " #addr_mode;                               \
-        printTestStartMsg( testName );                                                                       \
-        json const testCases = extractTestsFromJson( "tests/json/" filename );                               \
+        std::string       space = " ";                                                                       \
+        std::string const testName = #opcode_hex + space + #mnemonic + space + #addr_mode;                   \
+        TestStart( testName );                                                                               \
+        std::string jsonDir = std::string( TESTS_DIR ) + "/json/" + ( filename );                            \
+        json const  testCases = GetJsonData( jsonDir );                                                      \
         for ( const auto &testCase : testCases ) {                                                           \
             RunTestCase( testCase );                                                                         \
         }                                                                                                    \
-        printTestEndMsg( testName );                                                                         \
+        TestEnd( testName );                                                                                 \
     }
-// NOLINTEND(cppcoreguidelines-macro-usage)
-
-/*
-  Testing an opcode:
-  1. The opcode is implemented in the CPU
-  2. The JSON file exists in tests/json
-  3. Uncomment the corresponding test below
-  4. Build and run the tests
-
-To isolate a test from the CLI:
-  ./scripts/test.sh "CPUTestFixture.x00" # runs only the BRK test
-  ./scripts/test.sh "CPUTestFixture.x01" # runs only the ORA IndirectX test
-  ./scripts/test.sh "CPUTestFixture.x05" # runs only the ORA ZeroPage test
-
-  or:
-  ctest -R "CPUTestFixture.x00" # from the build directory
-
-To run all tests:
-  ./scripts/test.sh
-  or
-  ctest # from the build directory
-*/
-
-/* CPU_TEST( SAMPLE, JSON, SANITY_CHECK, "temp.json" ); */
 
 /*
 ################################
-||                            ||
 ||      Official Opcodes      ||
-||                            ||
 ################################
 */
 CPU_TEST( 00, BRK, Implied, "00.json" );
@@ -511,340 +420,8 @@ CPU_TEST( 2B, ANC, Immediate, "2b.json" );
 CPU_TEST( AB, LXA, Immediate, "ab.json" );
 CPU_TEST( CB, SBX, Immediate, "cb.json" );
 CPU_TEST( BB, LAS, AbsoluteY, "bb.json" );
-CPU_TEST( 8B, ANE, AbsoluteY, "8b.json" );
-/*
-################################################################
-||                                                            ||
-||                    Test Fixture Methods                    ||
-||                                                            ||
-################################################################
-*/
+CPU_TEST( 8B, ANE, Immediate, "8b.json" );
 
-void CPUTestFixture::RunTestCase( const json &testCase ) // NOLINT
-{
-    // Initialize CPU
-    bus.EnableJsonTestMode();
-    cpu.EnableJsonTestMode();
-    ppu.EnableJsonTestMode();
-    cpu.Reset();
-
-    LoadStateFromJson( testCase, "initial" );
-    std::string const initialState = GetCPUStateString( testCase, "initial" );
-    // Ensure loaded values match JSON values
-    EXPECT_EQ( cpu.GetProgramCounter(), static_cast<u16>( testCase["initial"]["pc"] ) );
-    EXPECT_EQ( cpu.GetAccumulator(), testCase["initial"]["a"] );
-    EXPECT_EQ( cpu.GetXRegister(), testCase["initial"]["x"] );
-    EXPECT_EQ( cpu.GetYRegister(), testCase["initial"]["y"] );
-    EXPECT_EQ( cpu.GetStackPointer(), testCase["initial"]["s"] );
-    EXPECT_EQ( cpu.GetStatusRegister(), testCase["initial"]["p"] );
-
-    for ( const auto &ramEntry : testCase["initial"]["ram"] ) {
-        uint16_t const address = ramEntry[0];
-        uint8_t const  value = ramEntry[1];
-        EXPECT_EQ( cpu.Read( address ), value );
-    }
-
-    // Fetch, decode, execute
-    cpu.DecodeExecute();
-
-    // Check final state
-    bool               testFailed = false; // Track if any test has failed
-    std::ostringstream errorMessages;      // Accumulate error messages
-                                           //
-    if ( cpu.GetProgramCounter() != static_cast<u16>( testCase["final"]["pc"] ) ) {
-        testFailed = true;
-        errorMessages << "PC ";
-    }
-    if ( cpu.GetAccumulator() != static_cast<u8>( testCase["final"]["a"] ) ) {
-        testFailed = true;
-        errorMessages << "A ";
-    }
-    if ( cpu.GetXRegister() != static_cast<u8>( testCase["final"]["x"] ) ) {
-        testFailed = true;
-        errorMessages << "X ";
-    }
-    if ( cpu.GetYRegister() != static_cast<u8>( testCase["final"]["y"] ) ) {
-        testFailed = true;
-        errorMessages << "Y ";
-    }
-    if ( cpu.GetStackPointer() != static_cast<u8>( testCase["final"]["s"] ) ) {
-        testFailed = true;
-        errorMessages << "S ";
-    }
-    if ( cpu.GetStatusRegister() != static_cast<u8>( testCase["final"]["p"] ) ) {
-        testFailed = true;
-        errorMessages << "P ";
-    }
-    if ( cpu.GetCycles() != testCase["cycles"].size() ) {
-        testFailed = true;
-        errorMessages << "Cycle count ";
-    }
-
-    for ( const auto &ramEntry : testCase["final"]["ram"] ) {
-        uint16_t const address = ramEntry[0];
-        uint8_t const  expectedValue = ramEntry[1];
-        uint8_t const  actualValue = cpu.Read( address );
-        if ( actualValue != expectedValue ) {
-            testFailed = true;
-            errorMessages << "RAM ";
-        }
-    }
-
-    std::string const finalState = GetCPUStateString( testCase, "final" );
-    // print initial and final state if there are any failures
-    if ( testFailed ) {
-        std::cout << "Test Case: " << testCase["name"] << '\n';
-        std::cout << "Failed: " << errorMessages.str() << '\n';
-        std::cout << initialState << '\n';
-        std::cout << finalState << '\n';
-        std::cout << '\n';
-        FAIL();
-    }
-}
-
-void CPUTestFixture::LoadStateFromJson( const json &jsonData, const std::string &state )
-{
-    /*
-     This function loads the CPU state from json data.
-     args:
-      jsonData: JSON data returned by extractTestsFromJson
-      state: "initial" or "final"
-    */
-    cpu.SetProgramCounter( jsonData[state]["pc"] );
-    cpu.SetAccumulator( jsonData[state]["a"] );
-    cpu.SetXRegister( jsonData[state]["x"] );
-    cpu.SetYRegister( jsonData[state]["y"] );
-    cpu.SetStackPointer( jsonData[state]["s"] );
-    cpu.SetStatusRegister( jsonData[state]["p"] );
-
-    // Load memory state from JSON
-    for ( const auto &ramEntry : jsonData[state]["ram"] ) {
-        uint16_t const address = ramEntry[0];
-        uint8_t const  value = ramEntry[1];
-        cpu.Write( address, value );
-    }
-}
-
-std::string CPUTestFixture::GetCPUStateString( const json &jsonData, const std::string &state ) const
-{
-    /*
-    This function provides formatted output for expected vs. actual CPU state values,
-    based on provided json data and actual CPU state.
-    Args:
-      jsonData: JSON data returned by extractTestsFromJson
-      state: "initial" or "final"
-    */
-
-    // Expected values
-    u16 const expectedPc = static_cast<u16>( jsonData[state]["pc"] );
-    u8 const  expectedA = jsonData[state]["a"];
-    u8 const  expectedX = jsonData[state]["x"];
-    u8 const  expectedY = jsonData[state]["y"];
-    u8 const  expectedS = jsonData[state]["s"];
-    u8 const  expectedP = jsonData[state]["p"];
-    u64 const expectedCycles = jsonData["cycles"].size();
-
-    // Actual values
-    u16 const actualPc = cpu.GetProgramCounter();
-    u8 const  actualA = cpu.GetAccumulator();
-    u8 const  actualX = cpu.GetXRegister();
-    u8 const  actualY = cpu.GetYRegister();
-    u8 const  actualS = cpu.GetStackPointer();
-    u8 const  actualP = cpu.GetStatusRegister();
-    u64 const actualCycles = cpu.GetCycles();
-
-    // Column Widths
-    constexpr int labelWidth = 8;
-    constexpr int valueWidth = 14;
-
-    // Use osstringstream to collect output
-    std::ostringstream output;
-
-    // Print header
-    output << "----------" << state << " State----------" << '\n';
-    output << std::left << std::setw( labelWidth ) << "" << std::setw( valueWidth ) << "EXPECTED"
-           << std::setw( valueWidth ) << "ACTUAL" << '\n';
-
-    // Function to format and print a line
-    auto printLine = [&]( const std::string &label, const uint64_t expected, const uint64_t actual ) {
-        auto toHexDecimalString = []( const uint64_t value, const int width ) {
-            std::stringstream strStream;
-            strStream << std::hex << std::uppercase << std::setw( width ) << std::setfill( '0' ) << value
-                      << " (" << std::dec << value << ")";
-            return strStream.str();
-        };
-
-        int width; // NOLINT
-        if ( expected > 0xFFFF || actual > 0xFFFF ) {
-            width = 8;
-        } else if ( expected > 0xFF || actual > 0xFF ) {
-            width = 4;
-        } else {
-            width = 2;
-        }
-
-        output << std::left << std::setw( labelWidth ) << label;
-        output << std::setw( valueWidth ) << toHexDecimalString( expected, width );
-        output << std::setw( valueWidth ) << toHexDecimalString( actual, width ) << '\n';
-    };
-
-    // Print registers
-    printLine( "pc:", expectedPc, actualPc );
-    printLine( "s:", expectedS, actualS );
-    printLine( "a:", expectedA, actualA );
-    printLine( "x:", expectedX, actualX );
-    printLine( "y:", expectedY, actualY );
-    printLine( "p:", expectedP, actualP );
-
-    if ( state == "final" ) {
-        output << std::left << std::setw( labelWidth ) << "cycles:";
-        output << std::setw( valueWidth ) << expectedCycles;
-        output << std::setw( valueWidth ) << actualCycles << '\n';
-    }
-
-    // Blank line and RAM header
-    output << '\n' << "RAM" << '\n';
-
-    // Print RAM entries
-    for ( const auto &ramEntry : jsonData[state]["ram"] ) {
-        uint16_t const address = ramEntry[0];
-        uint8_t const  expectedValue = ramEntry[1];
-        uint8_t const  actualValue = cpu.Read( address );
-
-        // Helper lambda to format values as "HEX (DECIMAL)"
-        auto formatValue = []( const uint8_t value ) {
-            std::ostringstream oss;
-            oss << std::hex << std::uppercase << std::setw( 2 ) << std::setfill( '0' )
-                << static_cast<int>( value ) << " (" << std::dec << static_cast<int>( value ) << ")";
-            return oss.str();
-        };
-
-        // Format address as hex only (no decimal for addresses)
-        std::ostringstream addressStream;
-        addressStream << std::hex << std::setw( 4 ) << std::setfill( '0' ) << address;
-
-        // Print formatted output
-        output << std::left << std::setw( labelWidth ) << addressStream.str();
-        output << std::setw( valueWidth ) << formatValue( expectedValue );
-        output << std::setw( valueWidth ) << formatValue( actualValue ) << '\n';
-    }
-
-    output << "--------------------------------" << '\n';
-    output << '\n';
-
-    // Return the accumulated string
-    return output.str();
-}
-/*
-########################################
-||    Expose private methods here     ||
-########################################
-*/
-
-void CPUTestFixture::SetFlags( const u8 flag )
-{
-    cpu.SetFlags( flag );
-}
-void CPUTestFixture::ClearFlags( const u8 flag )
-{
-    cpu.ClearFlags( flag );
-}
-bool CPUTestFixture::IsFlagSet( const u8 flag ) const
-{
-    return cpu.IsFlagSet( flag );
-}
-u8 CPUTestFixture::Read( const u16 address ) const
-{
-    return cpu.Read( address );
-}
-void CPUTestFixture::Write( const u16 address, const u8 data ) const
-{
-    cpu.Write( address, data );
-}
-u16 CPUTestFixture::IMM()
-{
-    return cpu.IMM();
-}
-u16 CPUTestFixture::ZPG()
-{
-    return cpu.ZPG();
-}
-u16 CPUTestFixture::ZPGX()
-{
-    return cpu.ZPGX();
-}
-u16 CPUTestFixture::ZPGY()
-{
-    return cpu.ZPGY();
-}
-u16 CPUTestFixture::ABS()
-{
-    return cpu.ABS();
-}
-u16 CPUTestFixture::ABSX()
-{
-    return cpu.ABSX();
-}
-u16 CPUTestFixture::ABSY()
-{
-    return cpu.ABSY();
-}
-u16 CPUTestFixture::IND()
-{
-    return cpu.IND();
-}
-u16 CPUTestFixture::INDX()
-{
-    return cpu.INDX();
-}
-u16 CPUTestFixture::INDY()
-{
-    return cpu.INDY();
-}
-u16 CPUTestFixture::REL()
-{
-    return cpu.REL();
-}
-
-/*
-################################################################
-||                                                            ||
-||                  General Helper Functions                  ||
-||                                                            ||
-################################################################
-*/
-
-auto extractTestsFromJson( const std::string &path ) -> json
-// Extracts test cases from a JSON file and returns them as a JSON object, with the
-// help of the nlohmann::json library.
-{
-    std::ifstream jsonFile( path );
-    if ( !jsonFile.is_open() ) {
-        throw std::runtime_error( "Could not open test file: " + path );
-    }
-    json testCases;
-    jsonFile >> testCases;
-
-    if ( !testCases.is_array() ) {
-        throw std::runtime_error( "Expected an array of test cases in JSON file" );
-    }
-    return testCases;
-}
-
-void printTestStartMsg( const std::string &testName )
-{
-    std::cout << '\n';
-    std::cout << "---------- " << testName << " Tests ---------" << '\n';
-}
-void printTestEndMsg( const std::string &testName )
-{
-    std::cout << "---------- " << testName << " Tests Complete ---------" << '\n';
-    std::cout << '\n';
-}
-
-// -----------------------------------------------------------------------------
-// -------------------------------- MAIN ---------------------------------------
-// -----------------------------------------------------------------------------
 auto main( int argc, char **argv ) -> int
 {
     testing::InitGoogleTest( &argc, argv );
