+++ tests/json/temp.json
@@ -1,41 +1,2 @@
-[
-  {
-    "name": "ad dd d7",
-    "initial": {
-      "pc": 54886,
-      "s": 164,
-      "a": 131,
-      "x": 34,
-      "y": 155,
-      "p": 103,
-      "ram": [
-        [54886, 173],
-        [54887, 221],
-        [54888, 215],
-        [55261, 168],
-        [54889, 162]
-      ]
-    },
-    "final": {
-      "pc": 54889,
-      "s": 164,
-      "a": 168,
-      "x": 34,
-      "y": 155,
-      "p": 229,
-      "ram": [
-        [54886, 173],
-        [54887, 221],
-        [54888, 215],
-        [54889, 162],
-        [55261, 168]
-      ]
-    },
-    "cycles": [
-      [54886, 173, "read"],
-      [54887, 221, "read"],
-      [54888, 215, "read"],
-      [55261, 168, "read"]
-    ]
-  }
-]
+
+{ "name": "11 1b 8f", "initial": { "pc": 19390, "s": 95, "a": 224, "x": 196, "y": 65, "p": 169, "ram": [ [19390, 17], [19391, 27], [19392, 143], [27, 206], [28, 193], [49423, 216], [49679, 62]]}, "final": { "pc": 19392, "s": 95, "a": 254, "x": 196, "y": 65, "p": 169, "ram": [ [27, 206], [28, 193], [19390, 17], [19391, 27], [19392, 143], [49423, 216], [49679, 62]]}, "cycles": [ [19390, 17, "read"], [19391, 27, "read"], [27, 206, "read"], [28, 193, "read"], [49423, 216, "read"], [49679, 62, "read"]] },
